added admin verify middleware

+++ Controllers/postController.js
@@ -1,10 +1,14 @@
 const Post = require("../Models/posts");
 const Comment = require("../Models/comments");
+const User=require('../Models/Users');
 
 const createPost = async (req, res) => {
-  const { title, userId } = req.body;
-  const post = new Post({ title, userId, publishDate: new Date() });
+  const { title } = req.body;
+  const post = new Post({ title, userId:req.id, publishDate: new Date() });
   await post.save();
+  const user=await User.findById(req.id);
+  user.postId=post.id;
+  await user.save();
   res.send({ message: "Post created successfully", post });
 };
 
@@ -31,15 +35,13 @@ const deletePostById = async (req, res) => {
 };
 
 const deleteAllPosts = async (req, res) => {
-  //needs a fix
-  await Post.deleteMany({});
+  await Post.deleteMany({userId:req.id});
   res.send({ message: "All posts deleted successfully" });
 };
 
 const getAllPostsByUser = async (req, res) => {
   console.log("helloworld");
-  const { userId } = req.params;
-  const posts = await Post.find({ userId: { $eq: userId } });
+  const posts = await Post.find({ userId: req.id });
   res.send({ message: "All posts retrieved successfully", posts });
 };
 
@@ -48,6 +50,7 @@ const deleteAllPostsByUser = async (req, res) => {
   await Post.deleteMany({ userId });
   res.send({ message: "All posts deleted successfully" });
 };
+
 const getAllCommentsByPost = async (req, res) => {
   const { postId } = req.params;
   const comments = await Comment.find({ postId });
@@ -63,5 +66,5 @@ module.exports = {
   deleteAllPosts,
   getAllPostsByUser,
   deleteAllPostsByUser,
-  getAllCommentsByPost,
+  getAllCommentsByPost
 };
